+++ src/routes/IndexRoutes.ts
@@ -1,13 +1,13 @@
 import { Router } from 'express'
 import createError from 'http-errors'
 import { Client } from 'pg'
-import TypeService from '../services/TypeService'
-import OwnerService from '../services/OwnerService'
-import TimeService from '../services/TimeService'
+import TypeService from '../controllers/TypeService'
+import OwnerService from '../controllers/OwnerService'
+import TimeService from '../controllers/TimeService'
 import mapTime from '../utils/mapTime'
 import { sidebarComponent } from '../components/sidebar/sidebar'
 import { authenticated, hasAccess } from '../utils/auth'
-import { RoleService } from '../services/RoleService'
+import { RoleService } from '../controllers/RoleService'
 
 export function IndexRoutes (pgClient: Client): Router {
   const timeService = new TimeService(pgClient)
