+++ src/routes/RolesRoutes.ts
@@ -3,7 +3,7 @@ import { Client } from 'pg'
 import { tableComponent } from '../components/table/table'
 import toTableArray from '../utils/tableArray'
 import { sidebarComponent } from '../components/sidebar/sidebar'
-import { RoleService } from '../services/RoleService'
+import { RoleService } from '../controllers/RoleService'
 import { hasAccess } from '../utils/auth'
 import { validateBody, validateQuery } from '../utils/validateQuery'
 import createHttpError from 'http-errors'
