+++ src/routes/StatsRoutes.ts
@@ -1,11 +1,11 @@
 import { Router } from 'express'
 import { groupBy } from '../utils/json'
 import { Client } from 'pg'
-import ProjectsService from '../services/ProjectService'
-import OwnerService from '../services/OwnerService'
+import ProjectsService from '../controllers/ProjectService'
+import OwnerService from '../controllers/OwnerService'
 import { sidebarComponent } from '../components/sidebar/sidebar'
 import { hasAccess } from '../utils/auth'
-import { RoleService } from '../services/RoleService'
+import { RoleService } from '../controllers/RoleService'
 import { owner } from '../models/owner'
 import { projectDetail } from '../models/projectDetail'
