+++ src/routes/TeamRoutes.ts
@@ -2,10 +2,10 @@ import { Router } from 'express'
 import mapTime from '../utils/mapTime'
 import { groupBy } from '../utils/json'
 import { Client } from 'pg'
-import TimeService from '../services/TimeService'
+import TimeService from '../controllers/TimeService'
 import { sidebarComponent } from '../components/sidebar/sidebar'
 import { hasAccess } from '../utils/auth'
-import { RoleService } from '../services/RoleService'
+import { RoleService } from '../controllers/RoleService'
 import createHttpError from 'http-errors'
 import { Converter } from 'showdown'
 import { teamTime } from '../models/time'
