+++ src/routes/UsersRoutes.ts
@@ -2,12 +2,12 @@ import { Router } from 'express'
 import { Client } from 'pg'
 import toTableArray from '../utils/tableArray'
 import { sidebarComponent } from '../components/sidebar/sidebar'
-import { UserService } from '../services/UserService'
+import { UserService } from '../controllers/UserService'
 import { tableComponent } from '../components/table/table'
 import { Parser } from 'json2csv'
 import { hasAccess } from '../utils/auth'
 import { validateBody } from '../utils/validateQuery'
-import { RoleService } from '../services/RoleService'
+import { RoleService } from '../controllers/RoleService'
 import createHttpError from 'http-errors'
 
 export function UsersRoutes(pgClient: Client): Router {
